requirements.txt
django
djangorestframework
celery
gunicorn
djoser
whitenoise
django-cors-headers
redis
postgres